+++ src/main/java/com/coniverse/dangjang/domain/healthMetric/dto/response/HealthMetricResponse.java
@@ -2,8 +2,8 @@ package com.coniverse.dangjang.domain.healthMetric.dto.response;
 
 import java.time.LocalDate;
 
-import com.coniverse.dangjang.domain.healthMetric.entity.HealthMetricCode;
-import com.coniverse.dangjang.domain.healthMetric.entity.HealthMetricType;
+import com.coniverse.dangjang.domain.healthMetric.enums.HealthMetricCode;
+import com.coniverse.dangjang.domain.healthMetric.enums.HealthMetricType;
 
 /**
  * 건강지표 response dto abstract class이다.
